Split the MLX scalar dispatch into basic and math submodules
Mirrors the layout of pytensor/scalar/ itself. Other backends keep this in one module, but MLX has no scipy.special and so hand-rolls every series, which makes the file grow in a way theirs do not.
