+++ pytensor/link/mlx/dispatch/scalar/__init__.py
@@ -1,0 +1,1 @@
+from pytensor.link.mlx.dispatch.scalar import basic, math
